requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_full_version <= '3.11.2'
attrs==23.1.0 ; python_version >= '3.7'
autobahn==23.6.2 ; python_version >= '3.9'
automat==22.10.0
cffi==1.16.0 ; python_version >= '3.8'
channels[daphne]==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.4 ; python_version >= '3.7'
daphne==4.0.0
django==4.2.6
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1 ; python_version >= '3.5'
msgpack==1.0.7 ; python_version >= '3.8'
packaging==23.2 ; python_version >= '3.7'
psycopg2-binary==2.9.9
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyopenssl==23.2.0
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
redis==5.0.1 ; python_version >= '3.7'
service-identity==23.1.0
setuptools==68.2.2 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
twisted[tls]==23.8.0 ; python_full_version >= '3.7.1'
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
uritemplate==4.1.1 ; python_version >= '3.6'
zope-interface==6.1 ; python_version >= '3.7'
